add check supportsAsyncDelete

+++ openmetadata-service/src/test/java/org/openmetadata/service/resources/EntityResourceTest.java
@@ -251,7 +251,7 @@ public abstract class EntityResourceTest<T extends EntityInterface, K extends Cr
   protected boolean supportsPatch = true;
   protected final boolean supportsSoftDelete;
   protected boolean supportsFieldsQueryParam = true;
-  protected boolean supportsAsyncHardDelete = true;
+  protected boolean supportsAsyncDelete = true;
   protected final boolean supportsEmptyDescription;
 
   // Special characters supported in the entity name
@@ -2200,6 +2200,9 @@ public abstract class EntityResourceTest<T extends EntityInterface, K extends Cr
   @Test
   @Execution(ExecutionMode.CONCURRENT)
   void delete_async_nonExistentEntity_404() {
+    if (!supportsAsyncDelete) {
+      return;
+    }
     assertResponse(
         () -> deleteEntityAsync(NON_EXISTENT_ENTITY, false, false, ADMIN_AUTH_HEADERS),
         NOT_FOUND,
@@ -2209,6 +2212,9 @@ public abstract class EntityResourceTest<T extends EntityInterface, K extends Cr
   @Test
   @Execution(ExecutionMode.CONCURRENT)
   void delete_async_entity_as_non_admin_401(TestInfo test) throws HttpResponseException {
+    if (!supportsAsyncDelete) {
+      return;
+    }
     // Create entity as admin
     K request = createRequest(getEntityName(test), "", "", null);
     T entity = createEntity(request, ADMIN_AUTH_HEADERS);
@@ -2223,7 +2229,7 @@ public abstract class EntityResourceTest<T extends EntityInterface, K extends Cr
   @Test
   @Execution(ExecutionMode.CONCURRENT)
   public void delete_async_with_recursive_hardDelete(TestInfo test) throws Exception {
-    if (!supportsAsyncHardDelete) {
+    if (!supportsAsyncDelete) {
       return;
     }
     K request = createRequest(getEntityName(test), "", "", null);
@@ -2236,6 +2242,25 @@ public abstract class EntityResourceTest<T extends EntityInterface, K extends Cr
     assertEquals(entity.getName(), deleteMessage.getEntityName());
     assertNull(deleteMessage.getError());
     assertEntityDeleted(entity.getId(), true);
+  }
+
+  @Test
+  @Execution(ExecutionMode.CONCURRENT)
+  void delete_async_soft_delete(TestInfo test) throws Exception {
+    if (!supportsAsyncDelete && !supportsSoftDelete) {
+      return;
+    }
+
+    K request = createRequest(getEntityName(test), "", "", null);
+    T entity = createEntity(request, ADMIN_AUTH_HEADERS);
+
+    // Test async soft delete
+    DeleteEntityMessage deleteMessage = receiveDeleteEntityMessage(entity.getId(), false, false);
+
+    assertEquals("COMPLETED", deleteMessage.getStatus());
+    assertEquals(entity.getName(), deleteMessage.getEntityName());
+    assertNull(deleteMessage.getError());
+    assertEntityDeleted(entity.getId(), false);
   }
 
   protected DeleteEntityMessage receiveDeleteEntityMessage(
@@ -2318,25 +2343,6 @@ public abstract class EntityResourceTest<T extends EntityInterface, K extends Cr
     }
 
     return deleteMessage;
-  }
-
-  @Test
-  @Execution(ExecutionMode.CONCURRENT)
-  void delete_async_soft_delete(TestInfo test) throws Exception {
-    if (!supportsSoftDelete) {
-      return;
-    }
-
-    K request = createRequest(getEntityName(test), "", "", null);
-    T entity = createEntity(request, ADMIN_AUTH_HEADERS);
-
-    // Test async soft delete
-    DeleteEntityMessage deleteMessage = receiveDeleteEntityMessage(entity.getId(), false, false);
-
-    assertEquals("COMPLETED", deleteMessage.getStatus());
-    assertEquals(entity.getName(), deleteMessage.getEntityName());
-    assertNull(deleteMessage.getError());
-    assertEntityDeleted(entity.getId(), false);
   }
 
   protected javax.ws.rs.core.Response deleteEntityAsync(
